Va cham :)

+++ src/uet/oop/bomberman/entities/Entity.java
@@ -4,17 +4,18 @@ import javafx.scene.Scene;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 import uet.oop.bomberman.graphics.Sprite;
+import uet.oop.bomberman.map.TileMap;
 
 public abstract class Entity {
     //Tọa độ X tính từ góc trái trên trong Canvas
-    protected int x;
+    protected float x;
 
     //Tọa độ Y tính từ góc trái trên trong Canvas
-    protected int y;
+    protected float y;
     protected Image img;
 
     //Khởi tạo đối tượng, chuyển từ tọa độ đơn vị sang tọa độ trong canvas
-    public Entity(int xUnit, int yUnit, Image img ) {
+    public Entity(float xUnit, float yUnit, Image img ) {
         this.x = xUnit * Sprite.SCALED_SIZE;
         this.y = yUnit * Sprite.SCALED_SIZE;
         this.img = img;
